+++ frontend/src/styles/modal.ts
@@ -5,7 +5,6 @@ export const ModalContainer = css`
     display: flex;
     flex-direction: column;
     border-radius: 20px;
-    margin: 130px 0 0 235px;
     background-color: #f1f1f1;
     box-shadow: 0 4px 4px rgba(0, 0, 0, 0.1), 0 4px 20px rgba(0, 0, 0, 0.1);
 `;
